requirements.txt
requests
azure-storage-blob
pandas
numpy
azure-keyvault-secrets
azure-identity
